Move unit tests to algorithm history tests
